+++ src/includes/classes/Utils/Shortcode.php
@@ -274,6 +274,8 @@ class Shortcode extends SCoreClasses\SCore\Base\Core
             'current_user_bought_product'     => '', // Product ID (or SKU) expression.
             'current_user_can_download'       => '', // Product ID (or SKU) expression.
 
+            'request_var' => '', // Request var expression.
+
             // Attribute modifiers.
             '_for_blog' => '0', // A specific blog ID.
             '_satisfy'  => 'all', // `any` or `all` (default).
@@ -427,6 +429,21 @@ class Shortcode extends SCoreClasses\SCore\Base\Core
                     if ($this->current_atts[$_att_key]) {
                         $this->appendConditions($this->simpleExpr($_att_key, function ($product_id_or_sku) {
                             return '('.(int) $this->Wp->is_woocommerce_active.' && '.$this->current_user_id.' ? '.a::class.'::wcCustomerCanDownload('.$this->current_user_id.', '.c::sQuote($product_id_or_sku).') : false)';
+                        }));
+                    }
+                    break;
+
+                /*
+                 * Other utilities.
+                 */
+
+                /*
+                 * `request_var="[expr]"`
+                 */
+                case 'request_var':
+                    if ($this->current_atts[$_att_key]) {
+                        $this->appendConditions($this->simpleExpr($_att_key, function ($key) {
+                            return a::class.'::requestVar('.c::sQuote($key).')';
                         }));
                     }
                     break;